Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
bcrypt==4.0.1
python-jose[cryptography]
requests
google-auth
python-multipart
email-validator
pydantic
psycopg2-binary
python-socketio
aiofiles